fix(rev-245.2): audit G3 — 32-bit-wrap parity in renderer blend/lighting/gradient math
- pix3d-3-03 (bug): TextureRaster shade-interpolation step
  (arg8-arg7)*divTable[var16]>>6 is a 32-bit product in Java (Pix3D.java:2053)
  that overflows on bright wide textured triangles and changes texel
  brightness via (arg7>>23)&0x1F; wrapped with int(int32(...)).
- pix3d-2-01 (latent): TextureTriangle's nine u/v/w gradient cross-products
  and their per-scanline accumulators are 32-bit in Java (Pix3D.java:1420-1456);
  typed the gradient locals int32 (Go int32 arithmetic wraps mod 2^32 exactly
  like Java int) and sign-extend with int(...) at the 24 TextureRaster calls.
- model-3-03 (latent): ApplyLighting dot product / divisor / sum and
  MulColourLightness product (Model.java:1416-1462) wrapped to int32 before
  the division and >>7.
- pix32-02 (latent): TransPlot blend sum (Pix32.java:384) wrapped so the
  arithmetic >>8 sign-extends the top byte exactly as Java when bit 31 set.
- pixfont-01 (latent): DrawMaskAlpha rgb seed and per-pixel blend
  (PixFont.java:424,437) wrapped the same way, including the final store.
- tone-envelope-01: NOT re-typed — tone.go carries a standing documented
  decision (document, don't re-type); added the 2026-06-04 re-confirmation
  note naming the weakest-bounded site.

+++ pkg/jagex2/dash3d/model/model.go
@@ -1386,25 +1386,25 @@ func (m *Model) ApplyLighting(arg0, arg1, arg2, arg3, arg4 int) {
 		if m.FaceInfo == nil {
 			var12 := m.FaceColour[i]
 			var10 := m.VertexNormal[var7]
-			var11 := arg0 + (arg2*var10.X+arg3*var10.Y+arg4*var10.Z)/(arg1*var10.W)
+			var11 := int(int32(arg0 + int(int32(arg2*var10.X+arg3*var10.Y+arg4*var10.Z))/int(int32(arg1*var10.W))))
 			m.FaceColourA[i] = MulColourLightness(var12, var11, 0)
 			var14 := m.VertexNormal[var8]
-			var16 := arg0 + (arg2*var14.X+arg3*var14.Y+arg4*var14.Z)/(arg1*var14.W)
+			var16 := int(int32(arg0 + int(int32(arg2*var14.X+arg3*var14.Y+arg4*var14.Z))/int(int32(arg1*var14.W))))
 			m.FaceColourB[i] = MulColourLightness(var12, var16, 0)
 			var15 := m.VertexNormal[var9]
-			var17 := arg0 + (arg2*var15.X+arg3*var15.Y+arg4*var15.Z)/(arg1*var15.W)
+			var17 := int(int32(arg0 + int(int32(arg2*var15.X+arg3*var15.Y+arg4*var15.Z))/int(int32(arg1*var15.W))))
 			m.FaceColourC[i] = MulColourLightness(var12, var17, 0)
 		} else if m.FaceInfo[i]&0x1 == 0 {
 			var12 := m.FaceColour[i]
 			var13 := m.FaceInfo[i]
 			var10 := m.VertexNormal[var7]
-			var11 := arg0 + (arg2*var10.X+arg3*var10.Y+arg4*var10.Z)/(arg1*var10.W)
+			var11 := int(int32(arg0 + int(int32(arg2*var10.X+arg3*var10.Y+arg4*var10.Z))/int(int32(arg1*var10.W))))
 			m.FaceColourA[i] = MulColourLightness(var12, var11, var13)
 			var10 = m.VertexNormal[var8]
-			var11 = arg0 + (arg2*var10.X+arg3*var10.Y+arg4*var10.Z)/(arg1*var10.W)
+			var11 = int(int32(arg0 + int(int32(arg2*var10.X+arg3*var10.Y+arg4*var10.Z))/int(int32(arg1*var10.W))))
 			m.FaceColourB[i] = MulColourLightness(var12, var11, var13)
 			var10 = m.VertexNormal[var9]
-			var11 = arg0 + (arg2*var10.X+arg3*var10.Y+arg4*var10.Z)/(arg1*var10.W)
+			var11 = int(int32(arg0 + int(int32(arg2*var10.X+arg3*var10.Y+arg4*var10.Z))/int(int32(arg1*var10.W))))
 			m.FaceColourC[i] = MulColourLightness(var12, var11, var13)
 		}
 	}
@@ -1431,7 +1431,8 @@ func MulColourLightness(arg0, arg1, arg2 int) int {
 		}
 		return 127 - arg1
 	}
-	arg1 = (arg1 * (arg0 & 0x7F)) >> 7
+	// Java: Model.java:1462 — 32-bit product wraps before the >>7 (audit model-3-03)
+	arg1 = int(int32(arg1*(arg0&0x7F))) >> 7
 	if arg1 < 2 {
 		arg1 = 2
 	} else if arg1 > 126 {
