+++ pkg/jagex2/graphics/pixfont/pixfont.go
@@ -485,7 +485,9 @@ func (p *PixFont) DrawCharAlpha(arg0 []int8, arg2, arg3, arg4, arg5, arg6, arg7 
 }
 
 func (p *PixFont) DrawMaskAlpha(arg0 int, arg1 int, arg2 int, arg3 []int, arg4 []int8, arg5 int, arg6 int, arg7 int, arg8 int, arg10 int) {
-	var17 := ((((arg10 & 0xFF00FF) * arg5) & 0xFF00FF00) + (((arg10 & 0xFF00) * arg5) & 0xFF0000)) >> 8
+	// Java: PixFont.java:424 — 32-bit blend sum; arithmetic >>8 sign-extends the
+	// top byte when bit 31 is set (audit pixfont-01)
+	var17 := int(int32((((arg10&0xFF00FF)*arg5)&0xFF00FF00)+(((arg10&0xFF00)*arg5)&0xFF0000))) >> 8
 	var15 := 256 - arg5
 	for i := -arg0; i < 0; i++ {
 		for j := -arg2; j < 0; j++ {
@@ -493,7 +495,8 @@ func (p *PixFont) DrawMaskAlpha(arg0 int, arg1 int, arg2 int, arg3 []int, arg4 [
 				arg1++
 			} else {
 				var14 := arg3[arg1]
-				arg3[arg1] = (((((var14 & 0xFF00FF) * var15) & 0xFF00FF00) + (((var14 & 0xFF00) * var15) & 0xFF0000)) >> 8) + var17
+				// Java: PixFont.java:437 — same 32-bit blend; final store wraps too
+				arg3[arg1] = int(int32((int(int32((((var14&0xFF00FF)*var15)&0xFF00FF00)+(((var14&0xFF00)*var15)&0xFF0000))) >> 8) + var17))
 				arg1++
 			}
 			arg6++
